Separated Kv17 from Kv15 in form.py
