+++ openebs/form.py
@@ -2,16 +2,13 @@ from builtins import object
 import logging
 from crispy_forms.bootstrap import AccordionGroup, Accordion
 from crispy_forms.helper import FormHelper
-from crispy_forms.layout import Submit, Layout, Field, HTML, Div, Hidden
+from crispy_forms.layout import Submit, Layout, Field, HTML, Div#, Hidden
 from django.utils.timezone import now
 import floppyforms.__future__ as forms
 from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext_lazy as _
-from kv1.models import Kv1Stop, Kv1Journey
-from kv15.enum import REASONTYPE, SUBREASONTYPE, ADVICETYPE, SUBADVICETYPE
-from openebs.models import Kv15Stopmessage, Kv15Scenario, Kv15ScenarioMessage, Kv17Change, get_end_service
-from openebs.models import Kv17JourneyChange
-from utils.time import get_operator_date
+from kv1.models import Kv1Stop
+from openebs.models import Kv15Stopmessage, Kv15Scenario, Kv15ScenarioMessage, get_end_service
 
 log = logging.getLogger('openebs.forms')
 
@@ -223,97 +220,6 @@ class Kv15ScenarioMessageForm(forms.ModelForm):
         )
 
 
-class Kv17ChangeForm(forms.ModelForm):
-    # This is duplication, but should work
-    reasontype = forms.ChoiceField(choices=REASONTYPE, label=_("Type oorzaak"), required=False)
-    subreasontype = forms.ChoiceField(choices=SUBREASONTYPE, label=_("Oorzaak"), required=False)
-    reasoncontent = forms.CharField(max_length=255, label=_("Uitleg oorzaak"), required=False,
-                                    widget=forms.Textarea(attrs={'cols': 40, 'rows': 4, 'class': 'col-lg-6'}))
-    advicetype = forms.ChoiceField(choices=ADVICETYPE, label=_("Type advies"), required=False)
-    subadvicetype = forms.ChoiceField(choices=SUBADVICETYPE, label=_("Advies"), required=False)
-    advicecontent = forms.CharField(max_length=255, label=_("Uitleg advies"), required=False,
-                                    widget=forms.Textarea(attrs={'cols': 40, 'rows': 4, 'class': 'col-lg-6'}))
-
-    def clean(self):
-        cleaned_data = super(Kv17ChangeForm, self).clean()
-        if 'journeys' not in self.data:
-            raise ValidationError(_("Een of meer geselecteerde ritten zijn ongeldig"))
-
-        valid_journeys = 0
-        for journey in self.data['journeys'].split(',')[0:-1]:
-            journey_qry = Kv1Journey.objects.filter(pk=journey, dates__date=get_operator_date())
-            if journey_qry.count() == 0:
-                raise ValidationError(_("Een of meer geselecteerde ritten zijn ongeldig"))
-            if Kv17Change.objects.filter(journey__pk=journey, line=journey_qry[0].line,
-                                         operatingday=get_operator_date()).count() != 0:
-                raise ValidationError(_("Een of meer geselecteerde ritten zijn al aangepast"))
-            valid_journeys += 1
-
-        if valid_journeys == 0:
-            raise ValidationError(_("Er zijn geen ritten geselecteerd om op te heffen"))
-
-        return cleaned_data
-
-    def save(self, force_insert=False, force_update=False, commit=True):
-        ''' Save each of the journeys in the model. This is a disaster, we return the XML
-        TODO: Figure out a better solution fo this! '''
-        xml_output = []
-        for journey in self.data['journeys'].split(',')[0:-1]:
-            qry = Kv1Journey.objects.filter(id=journey, dates__date=get_operator_date())
-            if qry.count() == 1:
-                self.instance.pk = None
-                self.instance.journey = qry[0]
-                self.instance.line = qry[0].line
-                self.instance.operatingday = get_operator_date()
-                self.instance.is_cancel = True
-
-                # Unfortunately, we can't place this any earlier, because we don't have the dataownercode there
-                if self.instance.journey.dataownercode == self.instance.dataownercode:
-                    self.instance.save()
-
-                    # Add details
-                    if self.data['reasontype'] != '0' or self.data['advicetype'] != '0':
-                        Kv17JourneyChange(change=self.instance, reasontype=self.data['reasontype'],
-                                          subreasontype=self.data['subreasontype'],
-                                          reasoncontent=self.data['reasoncontent'],
-                                          advicetype=self.data['advicetype'],
-                                          subadvicetype=self.data['subadvicetype'],
-                                          advicecontent=self.data['advicecontent']).save()
-
-                    xml_output.append(self.instance.to_xml())
-                else:
-                    log.error(
-                        "Oops! mismatch between dataownercode of line (%s) and of user (%s) when saving journey cancel" %
-                        (self.instance.journey.dataownercode, self.instance.dataownercode))
-            else:
-                log.error("Failed to find journey %s" % journey)
-
-        return xml_output
-
-    class Meta(object):
-        model = Kv17Change
-        exclude = ['dataownercode', 'operatingday', 'line', 'journey', 'is_recovered', 'reinforcement']
-
-    def __init__(self, *args, **kwargs):
-        super(Kv17ChangeForm, self).__init__(*args, **kwargs)
-        self.helper = FormHelper()
-        self.helper.form_tag = False
-        self.helper.layout = Layout(
-            Accordion(
-                AccordionGroup(_('Oorzaak'),
-                               'reasontype',
-                               'subreasontype',
-                               'reasoncontent'
-                               ),
-                AccordionGroup(_('Advies'),
-                               'advicetype',
-                               'subadvicetype',
-                               'advicecontent'
-                               )
-            )
-        )
-
-
 class PlanScenarioForm(forms.Form):
     messagestarttime = forms.DateTimeField(label=_("Begin"), initial=now)
     messageendtime = forms.DateTimeField(label=_("Einde"), initial=get_end_service)
@@ -334,21 +240,4 @@ class PlanScenarioForm(forms.Form):
                 Div(Field('messageendtime'), css_class="col-sm-6 col-lg-6"),
                 css_class="row"),
             Submit('submit', _("Plan alle berichten in"))
-        )
-
-
-class CancelLinesForm(forms.Form):
-    verify_ok = forms.BooleanField(initial=True, widget=forms.HiddenInput)
-
-    def clean_verify_ok(self):
-        if self.cleaned_data.get('verify_ok') is not True:
-            raise ValidationError(_("Je moet goedkeuring geven om alle lijnen op te heffen"))
-
-    def __init__(self, *args, **kwargs):
-        super(CancelLinesForm, self).__init__(*args, **kwargs)
-        self.helper = FormHelper()
-        self.helper.form_tag = 'journey_redbutton'
-        self.helper.layout = Layout(
-            Hidden('verify_ok', 'true'),
-            Submit('submit', _("Hef alle ritten op"), css_class="text-center btn-danger btn-tall col-sm-3 pull-right")
         )
